+++ crates/oxrdp-pdu/src/lib.rs
@@ -19,8 +19,10 @@ pub mod ber;
 mod codec;
 pub mod connect;
 mod cursor;
+pub mod domain_params;
 mod error;
 pub mod gcc;
+pub mod gcc_server;
 pub mod mcs;
 pub mod nego;
 pub mod send_data;
@@ -30,8 +32,10 @@ pub mod x224;
 pub use codec::{decode, encode_vec, Decode, Encode};
 pub use connect::{ConnectionConfirm, ConnectionRequest, NegotiationConfirm};
 pub use cursor::{ReadCursor, WriteCursor};
+pub use domain_params::DomainParameters;
 pub use error::{DecodeError, DecodeResult, EncodeError, EncodeResult};
 pub use gcc::{ChannelDef, ClientCoreData, ClientNetworkData, ClientSecurityData};
+pub use gcc_server::{ServerCoreData, ServerNetworkData};
 pub use mcs::{
     AttachUserConfirm, AttachUserRequest, ChannelJoinConfirm, ChannelJoinRequest,
     ErectDomainRequest, MCS_USERCHANNEL_BASE,